requirements.txt
Flask==1.1.4
google-api-python-client==1.7.11
google-cloud==0.34.0